requirements.txt
langchain==0.0.154
python-dotenv==1.0.0
discord-py-slash-command==4.2.1